backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
fastapi-mail==1.4.1
pydantic[email]==2.10.1
pydantic-settings==2.2.1
python-multipart==0.0.9 
mangum==0.10.0
gunicorn==21.2.0